# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Instagram-clone API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username or email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Exchange a refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Revoke the current refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "204":
          description: Logged out successfully

  # ── Users ─────────────────────────────────────────────────────────────────
...
  /users/me/close-friends:
    get:
      operationId: getCloseFriends
      tags: [users]
      summary: Get my close friends list
      responses:
        "200":
          description: Close friends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
    post:
      operationId: addCloseFriend
      tags: [users]
      summary: Add a user to close friends
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
              required: [userId]
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
  /users/me/close-friends/{friendId}:
    delete:
      operationId: removeCloseFriend
      tags: [users]
      summary: Remove a user from close friends
      parameters:
        - name: friendId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
...
  /explore/posts:
    get:
      operationId: getExplorePosts
      tags: [explore]
      summary: Get explore page posts (trending/recent)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Explore posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
  /search/users:
    get:
      operationId: searchUsers
      tags: [explore]
      summary: Search for users by username or name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"

  # ── Vault ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth ──────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
      required:
        - username
        - email
        - password
        - fullName
    LoginInput:
      type: object
      properties:
        identifier:
          type: string
          description: Username or email
        password:
          type: string
      required:
        - identifier
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: Short-lived JWT access token (15 minutes)
        refreshToken:
          type: string
          description: Long-lived refresh token (30 days)
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - refreshToken
        - user
    RefreshTokenInput:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    RefreshTokenResponse:
      type: object
      properties:
        token:
          type: string
          description: New short-lived JWT access token
      required:
        - token

    # ── Users ──────────────────────────────────────────────────────────────
    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        fullName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        postsCount:
          type: integer
        followersCount:
          type: integer
        followingCount:
          type: integer
        isFollowing:
          type: boolean
        isMe:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - fullName
        - postsCount
        - followersCount
        - followingCount
        - isFollowing
        - isMe
        - createdAt
    UserSummary:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        fullName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isFollowing:
          type: boolean
      required:
        - id
        - username
        - fullName
        - isFollowing
...
    Post:
      type: object
      properties:
        id:
          type: string
        caption:
          type: ["string", "null"]
        mediaUrl:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        audience:
          type: string
          enum: [everyone, close_friends]
        location:
          type: ["string", "null"]
        altText:
          type: ["string", "null"]
        commentsDisabled:
          type: boolean
        additionalMediaUrls:
          type: array
          items:
            type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLiked:
          type: boolean
        isSaved:
          type: boolean
        author:
          $ref: "#/components/schemas/UserSummary"
        createdAt:
          type: string
      required:
        - id
        - mediaUrl
        - mediaType
        - audience
        - commentsDisabled
        - additionalMediaUrls
        - likesCount
        - commentsCount
        - isLiked
        - isSaved
        - author
        - createdAt
    PostsPage:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        hasMore:
          type: boolean
        total:
          type: integer
      required:
        - posts
        - hasMore
        - total
...